added user space applications

+++ kpaxos/include/kernel_device.h
@@ -11,7 +11,7 @@ extern ssize_t kdev_read(struct file*, char*, size_t, loff_t*);
 extern ssize_t kdev_write(struct file*, const char*, size_t, loff_t*);
 extern int     kdevchar_init(int id, char* name);
 extern void    kdevchar_exit(void);
-extern void    kset_message(char* msg, size_t size, unsigned int iid);
+extern void    kset_message(char* msg, int size, unsigned int iid);
 extern void    kstop_device(void);
 extern struct file_operations fops;
